Split EvaluationReport.cs into sensible files
